extract completion tag function test helper
